@@ -3,11 +3,11 @@
 	"artifactName": "ngx-leaflet-markercluster",
 	"moduleName": "ngxLeafletMarkercluster",
 	"description": "Angular.io components for leaflet markercluster plugin",
-	"version": "3.0.0",
+	"version": "4.0.0",
 	"author": "Asymmetrik, Ltd.",
-	"copyright": "Copyright Asymmetrik, Ltd. 2007-2019 - All Rights Reserved.",
+	"copyright": "Copyright Asymmetrik, Ltd. 2007-2020 - All Rights Reserved.",
 	"license": "MIT",
-	"main": "dist/bundles/ngx-leaflet-markercluster.js",
+	"main": "dist/bundles/ngx-leaflet-markercluster.umd.js",
 	"module": "dist/index.js",
 	"jsnext:main": "dist/index.js",
 	"typings": "dist/index.d.ts",
@@ -21,10 +21,10 @@
 		"demo": "(webpack-dev-server & npm run watch)",
 		"predemo": "npm run lint",
 		"bundle": "npm run rollup && npm run uglify",
-		"compile": "ngc -p tsconfig-aot.json",
+		"compile": "ngc -p src/tsconfig-lib.json",
 		"lint": "tslint -c tslint.json -t stylish 'src/**/*.ts'",
 		"rollup": "rollup -c rollup.config.js",
-		"uglify": "uglifyjs --comments 'license' -o ./dist/bundles/ngx-leaflet-markercluster.min.js -- ./dist/bundles/ngx-leaflet-markercluster.js",
+		"uglify": "uglifyjs --comments -o ./dist/bundles/ngx-leaflet-markercluster.min.js -- ./dist/bundles/ngx-leaflet-markercluster.js",
 		"watch": " watch 'npm run lint' ./src"
 	},
 	"files": [
@@ -32,50 +32,54 @@
 		"src"
 	],
 	"peerDependencies": {
-		"@angular/core": ">=8",
-		"@angular/common": ">=8",
+		"@angular/core": ">=9",
+		"@angular/common": ">=9",
 		"leaflet": "1",
 		"leaflet.markercluster": "1",
-		"@asymmetrik/ngx-leaflet": "5"
+		"@asymmetrik/ngx-leaflet": "7",
+		"tslib": "1"
 	},
 	"devDependencies": {
-		"@angular/common": "8",
-		"@angular/compiler": "8",
-		"@angular/compiler-cli": "8",
-		"@angular/core": "8",
-		"@angular/platform-browser": "8",
-		"@angular/platform-browser-dynamic": "8",
-		"@asymmetrik/ngx-leaflet": "6",
+		"@angular/common": "9",
+		"@angular/compiler": "9",
+		"@angular/compiler-cli": "9",
+		"@angular/core": "9",
+		"@angular/platform-browser": "9",
+		"@angular/platform-browser-dynamic": "9",
+		"@asymmetrik/ngx-leaflet": "7",
 		"@types/leaflet": "1",
 		"@types/leaflet.markercluster": "1",
-		"@types/node": "12",
+		"@rollup/plugin-commonjs": "11",
+		"@rollup/plugin-node-resolve": "7",
+		"@types/node": "13",
 		"angular2-template-loader": "0.6",
 		"awesome-typescript-loader": "5",
 		"bootstrap": "4",
-		"core-js": "2.6",
+		"core-js": "3",
 		"css-loader": "3",
-		"file-loader": "4",
-		"html-loader": "0.5",
-		"html-webpack-plugin": "3",
+		"file-loader": "6",
+		"html-loader": "1",
+		"html-webpack-plugin": "4",
 		"leaflet": "1",
 		"leaflet.markercluster": "1",
 		"node-sass": "4",
 		"resolve-url-loader": "3",
-		"rollup": "1",
+		"rollup": "2",
 		"rxjs": "6",
-		"sass-loader": "7",
+		"sass-loader": "8",
 		"source-map-loader": "0.2",
-		"style-loader": "0.23",
+		"style-loader": "1",
 		"to-string-loader": "1.1",
-		"tslint": "5",
-		"typescript": "3.4",
+		"tslib": "1",
+		"tslint": "6",
+		"typescript": "3.8",
 		"uglify-js": "3",
-		"url-loader": "2",
+		"url-loader": "4",
 		"watch": "1",
 		"webpack": "4",
 		"webpack-cli": "3",
-		"webpack-dev-server": "^3.1.14",
-		"webpack-stats-plugin": "0.2",
-		"zone.js": "0.9"
+		"webpack-dev-server": "3",
+		"webpack-stats-plugin": "0.3",
+		"zone.js": "0.10"
 	}
 }
