@@ -6,8 +6,7 @@
  */
 
 // Polyfills
-import 'core-js/es6';
-import 'core-js/es7/reflect';
+import 'core-js';
 
 
 // Global Imports
